+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - refactor: improve RabbitMQ client
 - chore(cli): improve gen task command
 - chore(response): remove init resp and improve Error
-- chore(server): improve redis consumer server
+- chore: improve redis queue and add example
 
 ## v1.7.0
 - feat(http): can custom http status